Tempates & routes refactoring
