+++ ui/app/controllers/snow/apps/show.js
