--- ui/app/routes/dashboard.js
@@ -1,10 +1,0 @@
-import Ember from 'ember';
-import AuthenticatedRouteMixin from 'simple-auth/mixins/authenticated-route-mixin';
-
-export default Ember.Route.extend(AuthenticatedRouteMixin, {
-  model: function() {
-    return Ember.RSVP.hash({
-      apps: this.store.find('app')
-    });
-  }
-});
