+++ ui/app/routes/snow/dashboard.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  model: function() {
+    return Ember.RSVP.hash({
+      apps: this.store.find('app')
+    });
+  }
+});
