+++ ui/app/templates/snow/apps/_info.hbs
