+++ ui/app/templates/snow/apps/_install.hbs
