+++ ui/app/templates/snow/apps/edit.hbs
@@ -2,5 +2,5 @@
   <h1>Edit Application</h1>
 </div>
 
-{{partial "apps/_form"}}
+{{partial "snow/apps/_form"}}
