+++ ui/app/templates/snow/apps/new.hbs
@@ -2,4 +2,4 @@
   <h1>Register Application</h1>
 </div>
 
-{{partial "apps/_form"}}
+{{partial "snow/apps/_form"}}
