+++ ui/app/templates/snow/apps/show.hbs
@@ -12,11 +12,11 @@
     <div class="tab-content">
       <div role="tabpanel" class="tab-pane active" id="info">
         <div>&nbsp;</div>
-        {{partial "apps/info"}}
+        {{partial "snow/apps/info"}}
       </div>
       <div role="tabpanel" class="tab-pane" id="install">
         <div>&nbsp;</div>
-        {{partial "apps/install"}}
+        {{partial "snow/apps/install"}}
       </div>
     </div>
   </div>
@@ -25,6 +25,6 @@
   <div class="page-header">
     <h1>{{app.name}} - Installation</h1>
   </div>
-  {{partial "apps/install"}}
+  {{partial "snow/apps/install"}}
 {{/if}}
