+++ ui/app/templates/snow/settings.hbs
