+++ ui/tests/unit/controllers/snow/apps/new-test.js
@@ -3,12 +3,13 @@ import {
   test
 } from 'ember-qunit';
 
-moduleFor('route:apps/edit', 'AppsEditRoute', {
+moduleFor('controller:snow/apps/new', 'AppsNewController', {
   // Specify the other units that are required for this test.
   // needs: ['controller:foo']
 });
 
+// Replace this with your real tests.
 test('it exists', function() {
-  var route = this.subject();
-  ok(route);
+  var controller = this.subject();
+  ok(controller);
 });
