+++ ui/tests/unit/routes/snow-test.js
@@ -3,13 +3,12 @@ import {
   test
 } from 'ember-qunit';
 
-moduleFor('controller:dashboard', 'DashboardController', {
+moduleFor('route:snow', 'SnowRoute', {
   // Specify the other units that are required for this test.
   // needs: ['controller:foo']
 });
 
-// Replace this with your real tests.
 test('it exists', function() {
-  var controller = this.subject();
-  ok(controller);
+  var route = this.subject();
+  ok(route);
 });
