+++ ui/tests/unit/routes/snow/apps/edit-test.js
@@ -3,13 +3,12 @@ import {
   test
 } from 'ember-qunit';
 
-moduleFor('controller:apps/new', 'AppsNewController', {
+moduleFor('route:snow/apps/edit', 'AppsEditRoute', {
   // Specify the other units that are required for this test.
   // needs: ['controller:foo']
 });
 
-// Replace this with your real tests.
 test('it exists', function() {
-  var controller = this.subject();
-  ok(controller);
+  var route = this.subject();
+  ok(route);
 });
